Update UI,fix bugs
